fix: RcBlockHeaderResponse position
