+++ crates/server/src/handlers/rc/blocks/get_head_header.rs
@@ -11,9 +11,11 @@ use axum::{
     http::StatusCode,
     response::{IntoResponse, Response},
 };
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use serde_json::json;
 use thiserror::Error;
+
+use crate::handlers::blocks::common::RcBlockHeaderResponse;
 
 /// Query parameters for /rc/blocks/head/header endpoint
 #[derive(Debug, Deserialize)]
@@ -25,17 +27,6 @@ pub struct RcBlockHeadHeaderQueryParams {
 
 fn default_finalized() -> bool {
     true
-}
-
-/// Relay chain block header response
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RcBlockHeaderResponse {
-    pub parent_hash: String,
-    pub number: String,
-    pub state_root: String,
-    pub extrinsics_root: String,
-    pub digest: serde_json::Value,
 }
 
 /// Error types for /rc/blocks/head/header endpoint
